Fix position issue on edit + finish creation project

+++ src/Controller/Admin/ProjectCrudController.php
@@ -34,19 +34,57 @@ class ProjectCrudController extends AbstractCrudController
         return Project::class;
     }
 
-    public function createEntity(string $entityFqcn)
-    {
-        $product = new Project();
-        $product->setName("Test 112");
+    // public function createEntity(string $entityFqcn)
+    // {
+    //     $product = new Project();
+    //     $product->setName("Test 112");
 
-        return $product;
-    }
+    //     return $product;
+    // }
 
     public function persistEntity(EntityManagerInterface $em, $entityInstance): void
     {
         if (!$entityInstance instanceof Project) return;
 
-        $list_project_images = $entityInstance->getProjectImages();
+        $list_project_images = $entityInstance->getProjectImages()->toArray();
+
+        $files = parent::getContext()->getRequest()->files;
+        $list_images_uploaded = array_values($files->get('Project')['projectImages']);
+
+        $request = parent::getContext()->getRequest()->request;
+        $listImageToDelete = array_values($request->all('Project')["projectImages"]);
+
+        if (count($list_images_uploaded) === 0) {
+            $entityInstance->setIsOnline(false);
+            $entityInstance->setInBiography(false);
+        } else {
+            $slugger = new AsciiSlugger();
+
+            foreach ($list_project_images as $index => $value) {
+                $hasToBeDeleted = false;
+
+                $pos = $value->getPosition();
+                $value = $list_project_images[$pos];
+                $file = $list_images_uploaded[$index]["name"];
+
+                if (array_key_exists('delete', $listImageToDelete[$index])) {
+                    $hasToBeDeleted = $listImageToDelete[$index]["delete"] == "on";
+                }
+
+                if($hasToBeDeleted == true || is_null($file)) {
+                    $entityInstance->removeProjectImage($value);
+                } else {
+                    $uniqueid = uniqid();
+                    $newFilename = "{$slugger->slug(strtolower($entityInstance->getName()))}-{$uniqueid}.{$file->guessExtension()}";
+
+                    $value->setName($newFilename);
+                    $file->move(
+                        $this->getParameter('projects_images_directory'),
+                        $newFilename
+                    );
+                }
+            }
+        }
 
         parent::persistEntity($em, $entityInstance);
     }
@@ -56,16 +94,11 @@ class ProjectCrudController extends AbstractCrudController
         if (!$entityInstance instanceof Project) return;
 
         $list_project_images = $entityInstance->getProjectImages()->toArray();
-        // usort($list_project_images, function ($a, $b) {
-        //     return ($a->getPosition() > $b->getPosition()) ? 1 : -1;
-        // });
 
-        // dd($list_project_images);
-        
         $files = parent::getContext()->getRequest()->files;
         $list_images_uploaded = array_values($files->get('Project')['projectImages']);
-        $request = parent::getContext()->getRequest()->request;
         
+        $request = parent::getContext()->getRequest()->request;
         $listImageToDelete = array_values($request->all('Project')["projectImages"]);
         // dd($list_project_images,  $listImageToDelete, $list_images_uploaded);
 
@@ -78,25 +111,27 @@ class ProjectCrudController extends AbstractCrudController
             $entityInstance->setInBiography(false);
         } else {
             $slugger = new AsciiSlugger();
-
+            // dd($list_project_images);
             foreach ($list_project_images as $index => $value) {
-                $pos = $index;
+                $pos = $value->getPosition();
                 $value = $list_project_images[$pos];
-                $file = $list_images_uploaded[$pos]["name"];
+                $file = $list_images_uploaded[$index]["name"];
 
                 $hasToBeDeleted = false;
-                if (array_key_exists('delete', $listImageToDelete[$pos])) {
-                    $hasToBeDeleted = $listImageToDelete[$pos]["delete"] == "on";
+                if (array_key_exists('delete', $listImageToDelete[$index])) {
+                    $hasToBeDeleted = $listImageToDelete[$index]["delete"] == "on";
                 }
 
                 if (is_null($file)) {
-                    if ($hasToBeDeleted == true) {
+                    if ($hasToBeDeleted == true && !is_null($value->getName())) {
                         $value->removeUpload($this->getParameter('projects_images_directory'));
                         $qb = $em->createQueryBuilder()->delete("App\Entity\ProjectImage", "pi")->where('pi.id = :id')->setParameter("id", $value->getId())->getQuery();
                         $qb->execute();
                     } else if (is_null($value->getName())) {
                         $entityInstance->removeProjectImage($value);
                     }
+                } else if($hasToBeDeleted == true) {
+                    $entityInstance->removeProjectImage($value);
                 } else {
                     $oldFileName = $value->getName();
                     $uniqueid = uniqid();
@@ -142,8 +177,8 @@ class ProjectCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         $current_year = date('Y');
-        $list_years_raw = range($current_year + 1, 2010);
-
+        $list_years_raw = range($current_year + 2, 2010);
+        $list_years_raw["0000"] = -1;
         $list_years = new \stdClass();
         foreach ($list_years_raw as $key => $value) {
             $list_years->$value = $value;
@@ -155,12 +190,12 @@ class ProjectCrudController extends AbstractCrudController
 
         yield BooleanField::new('is_online', "Mettre en ligne")
             ->renderAsSwitch(false)->onlyOnForms()
-            ->setHelp("Ne sera pas mis en ligne en absence d'images");
+            ->setHelp("Ne sera pas mis en ligne s'il n'y a pas d'images liées");
         yield BooleanField::new('is_online', "Est en ligne")->renderAsSwitch(false)->hideOnForm();
-        yield BooleanField::new('is_online', "Est en ligne")
+        yield BooleanField::new('is_online', "Mettre en ligne")
             ->renderAsSwitch(true)
-            ->hideOnForm()
-            ->setFormTypeOptions(['disabled' => 'disabled', 'title' => "Ne sera pas pris en compte s'il n'a pas d'images"]);
+            ->hideOnForm();
+            // ->setFormTypeOptions(['disabled' =>true, 'title' => "Ne sera pas pris en compte s'il n'a pas d'images"]);
 
         yield BooleanField::new('in_biography', "Afficher la première image dans la biographie")
             ->renderAsSwitch(false)
@@ -170,6 +205,8 @@ class ProjectCrudController extends AbstractCrudController
             ->hideOnForm();
 
         yield ChoiceField::new('year', "Année")
+            ->setHelp("Choisir l'année -1 pour ne pas afficher de date")
+            ->setFormTypeOptions(['required' => false])
             ->setChoices((array) $list_years);
         yield CollectionField::new('projectImages', 'Images associées')
             ->setEntryType(ProjectImageType::class)
